rename actions/actionTypes
