--- src/redux/actions/FontFaceActions.js
@@ -1,8 +1,0 @@
-import { LOAD } from '../constants/FontFaceActionTypes.js';
-
-export function load(className) {
-  return {
-    type: LOAD,
-    className
-  };
-}
