--- src/redux/constants/audioPlayerActionTypes.js
@@ -1,12 +1,0 @@
-export const SET_USER_AGENT = '@@quran/audioplayer/SET_USER_AGENT';
-export const SET_CURRENT_FILE = '@@quran/audioplayer/SET_CURRENT_FILE';
-export const SET_CURRENT_WORD = '@@quran/audioplayer/SET_CURRENT_WORD';
-export const PLAY = '@@quran/audioplayer/PLAY';
-export const PAUSE = '@@quran/audioplayer/PAUSE';
-export const NEXT = '@@quran/audioplayer/NEXT';
-export const SET_AYAH = '@@quran/audioplayer/SET';
-export const PREVIOUS = '@@quran/audioplayer/PREVIOUS';
-export const SET_REPEAT = '@@quran/audioplayer/SET_REPEAT';
-export const TOGGLE_SCROLL = '@@quran/audioplayer/TOGGLE_SCROLL';
-export const BUILD_ON_CLIENT = '@@quran/audioplayer/BUILD_ON_CLIENT';
-export const UPDATE = '@@quran/audioplayer/UPDATE';
