--- src/redux/constants/ayahsActionTypes.js
@@ -1,7 +1,0 @@
-export const LOAD = '@@quran/ayahs/LOAD';
-export const LOAD_SUCCESS = '@@quran/ayahs/LOAD_SUCCESS';
-export const LOAD_FAIL = '@@quran/ayahs/LOAD_FAIL';
-export const CLEAR_CURRENT = '@@quran/ayahs/CLEAR_CURRENT';
-export const SET_CURRENT_AYAH = '@@quran/ayahs/SET_CURRENT_AYAH';
-export const SET_CURRENT_WORD = '@@quran/ayahs/SET_CURRENT_WORD';
-export const CLEAR_CURRENT_WORD = '@@quran/ayahs/CLEAR_CURRENT_WORD';
