--- src/redux/constants/fontFaceActionTypes.js
@@ -1,1 +1,0 @@
-export const LOAD = '@@quran/fontFaces/LOAD';
