--- src/redux/constants/optionsActionTypes.js
@@ -1,2 +1,0 @@
-export const TOGGLE_READING_MODE = '@@quran/options/TOGGLE_READING_MODE';
-export const SET_OPTION = '@@quran/options/SET_OPTION';
