--- src/redux/constants/searchActionTypes.js
@@ -1,3 +1,0 @@
-export const SEARCH = '@@quran/search/LOAD';
-export const SEARCH_SUCCESS = '@@quran/search/LOAD_SUCCESS';
-export const SEARCH_FAIL = '@@quran/search/LOAD_FAIL';
