+++ src/redux/modules/fontFaces.js
@@ -1,6 +1,6 @@
-import { LOAD_SUCCESS } from '../constants/ayahsActionTypes.js';
-import { SEARCH_SUCCESS } from '../constants/searchActionTypes.js';
-import { LOAD } from '../constants/fontFaceActionTypes.js';
+import { LOAD_SUCCESS } from '../constants/ayahs.js';
+import { SEARCH_SUCCESS } from '../constants/search.js';
+import { LOAD } from '../constants/fontFace.js';
 
 export default function reducer(state = {}, action = {}) {
   switch (action.type) {
